chore: Update getAnamnesisCtrl.js file and related handlers for medical history
